+++ src/app/App.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { AppRouter } from './router';
+import { AppRouter } from './AppRouter';
 
 export const App = () => {
   return <AppRouter />;
